zebra: make possible creating route with interface index as nexthop
actually, ipv4/ipv6 route entries are associated to MPLS, in the case
route have a defined nexthop IP. In the case of vrf route leak for
reaching a L3VPN, labels must be added.
Contrary to vrf-lite backend with a nexthop IP defined, netns vrf
backend will have an outgoing interface from the vrf perspective. So,
when creating the route-entry with label, we don't want to setup a
gateway IP, but use an output interface index. The zebra changes permit
to set such combination, by allowing to pass 0 byte length prefixes.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>