bgp, pimd, zebra, lib: add vrf_route in zapi messages for nexthop tracking
a new identifier is added in the exchanged data for nexthop tracking.
This identifier will be the requester of the nexthop entry. When a
nexthop validation is requested, the user gives not only the target vrf
id where the nexthop is supposed to be, but also the source vrf id where
the prefix requesting the nexthop is located.
Both vrf_id information will be used to determine if a vrf route leak
entry is possible.
For instance, if the vrf backend is netns, then the source vrf will be
used to check if it is possible to reach target vrf by using cross vrf
interfaces. The routing information returned by nexthop tracker may
change, to take into account that new path to reach remote vrf.
In the case, the vrf backend is vrf-lite, the nexthop tracking will
identify this, and will return the same routing information as before.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>