zebra: modify route tracking with tuple index(nexthop, vrf source, vrf target)
The route tracker module is modified. Before, route tracker was storing
nexthop information based on nexthop IP address, and target vrf. Now,the
route tracker module is being added the source vrf as additional index.
This extension should be used for vrf route leaking. For instance, this
extension permits to resolve a route leak entry for vrf with netns
backends. By taking into account the source vrf, the returned routing
entry will contain the cross-vrf interface that should be used when a
nexthop has to be solved between two vrfs. The routing information
returned will not change in the case of vrf lite.
Also, because zebra rnh entries for a given nexthop is made up of a list of
entries, the list needs to be parsed, so as to properly clean each
nexthop cache entry. Then, the list will be flushed, if necessary.

PR=61803
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Alain Ritoux <alain.ritoux@6wind.com>