zebra: create route nexthop list per nexthop, to facilitate leaking
each route node found will have a list of route nexthop entry structure.
Each structure will be indexed with the vrf_id of the request. This will
facilitate to know if and how vrf route leaking will be done, in the
case a nexthop has to be reached by a separate vrf, especially in the
non vrf-lite implementation.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>