zebra: provision an interface index in route nexthop structure
This interface index will indicate for a vrf which interface to use when
doing vrf route leak, with netns based vrfs. This field is used locally
in zebra. When nht request for vrf route leak is done, show ip nht displays
some information more about which is the source vrf of the nht entry, and
which output interface it uses.

PR=61803
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Alain Ritoux <alain.ritoux@6wind.com>