zebra: provision an interface index in route nexthop structure
This interface index will indicate for a vrf which interface to use when
doing vrf route leak, with netns based vrfs. This field is used locally
in zebra, and will be communicated later with zapi interface.
When nht request for vrf route leak is done, show ip nht displays some
information more about which is the source vrf of the nht entry, and
which output interface it uses.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>