zebra: add recursive rnh information
when rnh applies to vrf route leak, and a nexthop entry has been
validated, then for a separate vrf, the nexthop may not be directly
reachable. For that, zebra returns the interface index where traffic
should be redirected to reach that nexthop entry.
That interface index is encapsulated in resolved route_entry, in the
nexthop resolved field. This permits keeping the original route_entry
information, and also permits adding the specific information about
crossing the borders.

PR=61803
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Alain Ritoux <alain.ritoux@6wind.com>