bgpd: modify bgp nht, add vrf_id source of the request
The bgp nexthop cache is being added the vrf_id source. This is the
origin vrf that requested to know the nexthop entry for the defined
target vrf.
This tuple will be used especially for vrf route leak cases. Actually,
zebra also handles the same information via nexthop tracker. That will
permit customise the response for each request BGP performs. That is to
say that it will not change anything if the vrf backend is vrf lite.
However, if the vrf backend is netns, an additional information
including the cross vrf to use will be embedded in the response. That
interface will be known from the original source vrf_id.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>