bgpd: create a label allocation for mpls route leaks with netns vrfs
A new mpls family of labels is created: LP_TYPE_VRF_VETH.
Crossing vrf netns by using an interface requires one hop more.
When doing a MPLS route leak from vpn to vrf X, reaching that
remote VPN is done by using the MPLS label from the entry, but also
appending the outside label ( because it is a recursive route) that is
needed to link vpns together ( LDP for instance). The assumption here,
is that there is no switching, nor routing after that action.
However, when we are on a vrf X with netns based vrfs, we have one hop
more. Thus, we must find a way in default vrf to switch incoming labeled
packet from one defined label to the vpn label. This way, the resulting
action will be the same from the result perspective.

The label allocation is stored in BGP. a cache list of mpls labels is
stored in the BGP nexthop cache entry. The allocation and disallocation
will be done, based on the tuple (remote vpn label, remote IP aka
nexthop).
In case a cross vrf needs to use an interface, a new label must be
retrieved. that callback is used for that.

PR=61803
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Alain Ritoux <alain.ritoux@6wind.com>