ripd: fix operational-data northbound callbacks
The get_next() and lookup_entry() callbacks should return structures of
the same type. This is because the returned list entry is passed to the
get_elem() callbacks, which must assume the 'list_entry' parameter is
of a single type.

This wasn't a problem before because we'd always call lookup_entry()
after get_next(), but it was realized this isn't necessary when iterating
sequentially over a YANG list. Now we need to ensure that the get_elem()
callbacks work with list entries returned by either get_next() or
lookup_entry().

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>