build: fix detection of the confd binary
The confd binary must be searched in the path given by the user and not
in the system standard directories. This is necessary because, the way
confd is distributed, multiple installations of confd (from different
versions) are possible. Users must choose at configure time which confd
installation they want to use.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>