bgpd: Fixed bgp daemon crash when a L2VNI is unconfigured
A L2VNI has been configured with export and import route target.
When a VNI is unconfigured it deletes all of its import and export
route-targets. There is a export route-target link list and import
route-target linked list per VNI. There are two loops in the deletion code.
In the 1st loop the route-target was deleted and freed, but list node
was not deleted and also the data of the node wasn't reset.
The second loop tries to delete and free an RT which was already freed
crashing the bgp daemon.

Signed-off-by: karamalla@vmware.com