defaults: set bgp default timer for datacenters to 1 second
Moving default bgp connection timer from 10 to 1 second.
The use case happens in a datacenter ( the SDN controller drives the
FRR). The BGP daemon tries to connect to the various DC-GWs. By default,
there is a 10 second gap between each retry neighborship establishment.
To detect a failure and take action, 10 second may be too much. That is
why a proposal of changing the value to 1 second by default is done. As
the settings are self to datacenter usage, the proposal is to hard set
this value.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>