zebra: Ebgp neighbor stuck in active state if multihop cli given before the update-source #3264
What is happening here is, when we try to create a BGP session b/w two neighbour, we open two sockets per neighbour.   Out of theses two sockets (which BGP opens per peer in the initial stage to bring up the session i.e one socket for sending connect request to the neighbour and the other is the listening socket, which would be waiting for the connect request from the neighbour.), the connect socket is kind of closed for this peer and the listening socket path is working fine.
So if a new connection request doesn’t arrive to either of this two neighbour, this session b/w these two peers would never come up. It’s a kind of deadlock where either of the peer is waiting on the other to initiate the connection.

Zebra was not responding to re-registration messages from bgp due to  which the next hop (which was reset in BGP) was not getting resolved. So bgp was not preceding ahead with this next hop and there was a deadlock
created for this next hop.  If similar scenario happens in the neighbour FRR router. As a result of that both the neighbour would NOT try to connect each other and will depend on each other to initiate the connection ,which is never going to happen
if both the neighbours are having this same bug, it’s a deadlock.

So, in-order to avoid this deadlock Zebra would respond to the client registration request, even if next-hop is already registered. This change is needed as bgp's neighbour machine, for this particular neighbour goes into dead lock state, and never send
any connect request to this neighbour.