zebra: update nexthop tracking upon interface status change
interface status change may signify that a virtual ethernet used to
cross vrf borders is up or down. then, as nexthop tracking is handling
vrf route leaking, it needs to be informed of those updates.
For instance, if the cross vrf interface changes to down or up, or the
remote is not clearly identified, then nexthop tracking.

PR=61803
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Alain Ritoux <alain.ritoux@6wind.com>