lib,zebra: add a new nexthop informational NEXTHOP_FLAG_INFO_ONLY
- Increase the nexthop flags value from 8bit to 16 bits.
- A new value is appended in nexthop flag : NEXTHOP_FLAG_INFO_ONLY.
This flag is used when zebra receives a request to add/delete a route
entry with nexthops crossing vrfs.

this flag avoids injecting route entries into the kernel FIB; and in the
same time, this information is useful when one wants to know what is the
path that will take the packet to reach such prefix. As you can see, the
flag is set when receiving connected routes ( nexthop interface index),
and on other cases ( nexthop ipv4/v6 with/without ifindex).

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>