lib: add new route entry flag ZEBRA_FLAG_CROSS_VRF_IFACE
the new flag ZEBRA_FLAG_CROSS_VRF_IFACE presence on route entry permits
to detect that this is a vrf route leak that is using an interface to
cross the borders between two vrfs. This is the case when using netns
based vrfs. This flag is shared between clients and zebra daemon in zapi
parameters.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>