bgpd: support for recursive routes with route leak with vrf netns
route leak originally does not require extra information since with
vrf-lite, the information is passed to the kernel as is. The recursive
information was kept also. In order to keep recursive information, the
original bgp information must be kept. But in order to make zebra
understand that a new path must be chosen, a new nexthop is inserted.
That new nexthop is the cross vrf interface to reach.
Also, in order to give that information to zebra, the new route entry
flag ZEBRA_FLAG_CROSS_VRF_IFACE is added to route entry structure, so
that one can remember that the first entry is specific and is used to
cross borders by using an interface.
In the case where more than one entry is available, a route entry is
inserted each time necessary. This is true in case ECMP to vpn, or ECMP
between 2 vrfs.

PR=61803
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Alain Ritoux <alain.ritoux@6wind.com>