zebra: permit installing ecmp entries with mpls label different only
when receiving multipath vrf route leaks, in the case the vrf backend is
a network namespace, then the nexthop kinds to use are nexthop ifindex
kind. because the two vrf route come from vpn, the same interface will
be used. that makes not possible for zebra to distinguish if the two
nexthops are the same or not. previously, the nexthop was marked as
duplicate, then it was not installed into the FIB. To distinguish it,
the labels are looked up and compared between nexthops. So that
multipath entries can be installed.

PR=61803
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Acked-by: Alain Ritoux <alain.ritoux@6wind.com>