FRRouting Release 6.0.2
This is a maintenance release for a security (Denial of Service) issue
found with the expermimental BGP VNC Attribute (255).
Other experiments using the same attribute may bring down BGP peers
if they are malformed for the VNC parsing. (As required by RFC).
The update disables the parsing of the VNC attribute by default
to avoid this.

Beside this change, this release is similar to 6.0.1

Major Changes since 6.0 (already in 6.0.1)
 - New Debian Package structure without backports
 - Fix Memory Leak on FreeBSD
 - Fix BGP unnumbered peer setup
 - Various fixes for label manager
 - Various other fixes

Signed-off-by: Martin Winter <mwinter@opensourcerouting.org>